requirements.txt
python-binance==1.0.19
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
rich==13.8.1
backoff==2.2.1
orjson==3.10.7